backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain_community
youtube-transcript-api
faiss-cpu
sentence-transformers
python-dotenv
pydantic